adds hook feature testing in app.js, new file structure, copy of old app.js in app2.js
